test: consolidate title-special-chars tests into transform.test.ts
Move the backtick and \$-escape regression tests into the existing
transform.test.ts file to match the project convention of grouping
related tests in shared files.
